+++ app/composables/render/idb.ts
@@ -35,8 +35,6 @@ export type IDBNavigraphProcedures = {
 class VatsimRadarDB extends Dexie {
     data!: Table<VatSpyAPIData | SimAwareAPIData | VatglassesAPIData | IDBAirlinesData, string>;
 
-    vatspy!: Table<Feature<MultiPolygon, VatSpyDataProperties>[] | string, string>;
-
     vatglasses!: Table<VatglassesPosition[] | VatglassesData[string] | string, string>;
 
     simaware!: Table<SimAwareDataFeature[] | string, string>;
@@ -57,10 +55,9 @@ export async function initClientDB() {
     indexedDB.deleteDatabase('vatsim-radar');
     const db = new VatsimRadarDB('vatsim-radar-db');
 
-    db.version(5)
+    db.version(6)
         .stores({
             data: '',
-            vatspy: '',
             vatglasses: '',
             simaware: '',
             airlines: '',