+++ app/utils/shared/runway-detection.ts
@@ -11,6 +11,7 @@ const common = [
     'LDG AND DEPTG',
     'LDG AND DEPTG RWYS',
     'ARRS AND DEPS',
+    'RUNWAY IN USE',
 ];
 
 const depMarkers = [
@@ -56,7 +57,7 @@ const arrMarkers = [
 
 const breakTentative = ['TLS', 'TRL', 'TRANSITION LEVEL', 'ARRS.', 'FOR ARRS. RWY'];
 
-const runwayRegex = /^(RWY)?(?<runway>\d{2}) ?(?<postfix>R|L|C|RIGHT|LEFT|CENTER)?(,|\.)?$/;
+const runwayRegex = /^(RWY)?(?<runway>\d{2}) ?(?<postfix>R|L|C|RIGHT|LEFT|CENTER)?(,|\.)+?$/;
 const auRegex = /\[RWY] (\d{2} ?(R|L|C|RIGHT|LEFT|CENTER)?)+$/;
 const auRegexCommon = /\[RWY] (\d{2} ?(R|L|C|RIGHT|LEFT|CENTER)?) ([\[+])/;
 
@@ -168,3 +169,6 @@ export function getActiveRunways(atis: string | string[]): { departure: string[]
         arrival: Array.from(arrRunways),
     };
 }
+
+// @ts-expect-error no types for this properity
+if (typeof window !== 'undefined') window.getActiveRunways = getActiveRunways;